Use the right constant pool for primitive/array
A null pointer as the constant pool in the method meta data will cause
a crash in gc during stack walk. The constant pool for a
newInstanceImplThunk will be the constant pool of the class to be
instantiated. If a primitive or an array is asked to be instantiated
through Class.newInstance, an exception is expected instead of a crash.
Since primitive and array classes don't have constant pool, this change
uses the constant pool of java/lang/Class for their newInstanceImplThunk
to fix the crash.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>