Fix Metadata Reclamation Bug
When the metadata is reclaimed (due to a recompilation of a method), the
previous and next metadatas in the linked list are are updated in
updateExceptionTableLinkedList. However, there's a corner case that
wasn't handled. If the metadata being reclaimed is the first in the
list, then the pointer to it (either off the classloader or the class
(if anonymous)) was not updated to point to the new stub metadata. This
commit rectifies this.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>