Reduce number of remote queries for VM_getBaseComponentClass
The solution is to cache the required information for each J9Class.
Since we already have the following hashtable
PersistentUnorderedMap<J9Class*, ClassInfo> _romClassMap;
in the clientSessionData, the job becomes just to add more fields
to ClassInfo and more parameters to some relevant functions.
The required information should be sent together with the compilation
request.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>