Improve start-up time when not using AOT
There are two changes implemented by this commit:
1) Reduce the invocation counts for methods belonging to bootstrap classes
2) Downgrade to cold compilations performed during start-up phase

The motivation for change (1) is to get out of interpreter sooner (the
invocation count is the number of times an interpreted method needs to
be invoked to be JIT compiled). The main problem is that a lower
invocation count reduces the number of interpreter profiling samples
which in turn may affect throughput. One way around this issue is to restrict
this change to methods belonging to bootstrap classes that are executed a
lot during start-up but to a lesser degree during stready state.

The motivation for change (2) is that, with a lower invocation count more
methods are going to be compiled and the extra compilation overhead may
affect start-up time. Thus, we need to reduce the compilation overhead
by using lower optimization levels (i.e. "cold"). Methods compiled at
"cold" will be upgraded through the GCR mechanism if they are executed
enough times.

There are two new options to control change (1)
-Xjit:bcountForBootstrapMethods=<nnn>   // for methods with loops
-Xjit:countForBootstrapMethods=<nnn>    // for methods without loops
and one new option to control change (2)
-Xjit:queueSizeThresholdToDowngradeOptLevelDuringStartup=<nnn>

Performance experiments on x86 showed good startup improvements for
Liberty Application Server (6-7%) and no throughput regressions for a
few Liberty based benchmarks.

In this commit the feature will be disabled to allow for more performance
testing on other platforms.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>