Separate address node of BCDCHK and fix data type
BCDCHK is a node that can GC-and-Return. As a result of this,
LocalCSE does not common BCDCHK's address node with any address
node after the BCDCHK.

When the BCDCHK produces a PackedDecimal type, it'll be followed by a
pdstorei node, which has the same address node as the BCDCHK node.
If this BCDCHK and it's following pdstorei share the same
address node, then no address node after the pdstorei could be commoned
with the pdstorei's address node because BCDCHK prevents this from
happening. Failing to common such address nodes can result in unnecessary
load-after-store instruction sequence (e.g. VLRL after VSTRL). This is
redundant and inefficient.

Separating BCDCHK's and pdstorei's address nodes makes the aforementioned
address node commoning possible and opens up opportunities for packed
decimal copy propagations.

Also, fix the symbol reference data type for the pdstorei node of
I2PD and PD-shift-left. The stored type should be TR::PackedDecimal
instead of TR::Int8.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>