Preserved original pointer in getCustomFFIType
The getIntFromLayout and getStructFFIType APIs take a reference to the
array returned by copyStringToUTF8WithMemAlloc and they modify the
pointer value. Subsequent calls to j9mem_free_memory will fail due to
metadata not being present.

We preserve the original array pointer and pass a temporary to the
aforementioned APIs to avoid these problems.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>