Fixed function call using implict "this"
Extensible classes should not use implicit "this" for member
function calls. The compiler needs to find  member functions
in a bottom up approach at build time to ensure the most
concrete method implementation is called first. To ensure that
always happens, member function calls in an extensible class
should be prefixed by 'self()->'.

This problem was detected by OMRChecker, and is a
prerequisite for enabling the OMRChecker linting in
Travis CI builds.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>