Fix devirtualization issues with IlGeneratorMethodDetails
When storing a particular variety of IlGeneratorMethodDetails into
MethodToBeCompiled, the code uses operator = which transmutes the existing
generic IlGeneratorMethodDetails (the base class) into a derived class
similar to the parameter that is passed in, by using placement new.
The expectation is that any call off the generic IlGeneratorMethodDetails
object stored in MethodToBeCompiled is going to be a virtual call.
However, the C++ compiler is allowed to devirtualize calls and if that
occurs, the JIT is going to invoke the implementation in the generic
IlGeneratorMethodDetails.

The solution introduced by this commit is to use an object factory
that builds a specific IlGeneratorMethodDetails at a designated address and
returns a pointer to that address which is then stored in MethodToBeCompiled.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>