Fix getStringUTFRegion not copying string argument correctly (#1710)
Fix getStringUTFRegion not copying string argument correctly

This JNI API should copy Unicode characters from offset start into str
for len number of characters. This is currently not happening as the
copyStringToUTF8WithMemAlloc will simply copy in the entire string.
This is undesirable.

We extend the copyStringToUTF8Helper API to handle offsets and lengths
as parameters. We make use of these paramters in the getStringUTFRegion
API to correctly copy the desired Unicode characters.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>