Remove unnecessary loads in deduplicateStrings
Currently the preferred backing array reference is loaded twice: once as
an integer, and again as a reference. Because of the differing types,
the JIT can't common these loads. This commit adds JITHelpers.acmplt so
that deduplicateStrings can use an address comparison instead of an
integer comparison, which allows us to use the backing array references
that have already been loaded by the caller. The JIT compiler will
replace the call to JITHelpers.acmplt with an acmplt IL node during
Unsafe fast pathing.

Also skip the ((s1.count ^ s2.count) >= 0) test unless

    enableCompression && null != compressionFlag

Otherwise counts are always nonnegative. In particular, when compression
is disabled, this test can be eliminated entirely.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>