Limit the invokestatic/special CP consistency check to JDK 9
The checks, added in PR #1167, are required for JDK 9.  We've had
reports of applications failing due to these ICCEs when running on
JDK 8.

This is a stop-gap solution to allow the JDK8 applications to continue
to run successfully while we get to the root cause on JDK8 as the
CP types look correct at first glance.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>