Fix toLowerCase combining above in Lithuanian
The toLowerCase API is not converting Lithuanian characters correctly.
The Unicode standard has updated the combining above code points to a
"Canonical Combining Above" class [1] where as the table used to be
just "Above" [3]. There were new code points introduced in [1] which
need to be properly updated in our implementations.

In addition the tests for lower casing Lithuanian are refactored and
partitioned into Java 8 and Java 9 tests since each Java version
supports a different level of the Unicode standard. The Java 8 test
has been updated in accordance to the updates to toLowerCase and
the Java 9 tests have been updated in accordance to [2].

[1] https://www.unicode.org/Public/6.2.0/ucd/extracted/DerivedCombiningClass.txt
[2] https://www.unicode.org/Public/8.0.0/ucd/extracted/DerivedCombiningClass.txt
[3] https://www.unicode.org/Public/4.0-Update/extracted/DerivedCombiningClass-4.0.0.txt

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>