J9 DDR generator file fixes and legacy mode
Fix index out of bounds in PointerGenerator.java. Code assumed there
would always be a number after '_v', which is erroneous if there is a
field with name ending like that.

Omit anonymous fields on aix since they start with #, an illegal
character, and aren't used in DDR.

Added legacy mode, where it checks for the suffix '_v#' for legacy DDR,
and it doesn't for new DDR since the new DDR does not generate the
offset suffixes for fields.

Change old world configs to use legacy mode in PointerGenerator and
StructureStubGenerator.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>