CMake fix bug in m4 generated files
When running in parallel, make can try generating 'generated.c' multiple
times in parallel. Causing unpredictable results. This is because each
target which uses a generated source file (from the same directory
where it was generate), gets its own rule on how to generate that file.

To work work arround this behaviour, the generation step is moved to the
top level CMakeLists. In order to make sure that the generated file exists
before compiling, we make j9vm_interface depend on its generation.

Signed-off-by: Devin Nakamura <devinn@ca.ibm.com>