Fix inlining of Unsafe.monitorEnter/Exit()
The JIT would use DirectToJNI for Unsafe.monitorEnter
and Unsafe.monitorExit. This would result in loadaddr
childern for the call. When inlining these calls we
use the 2nd child of the call as the monent/monexit
child, resulting an a stack reference being used as
if it was a object reference. This fix will convert
the loadaddr back to a aload when DirectToJNI was used

Signed-off-by: Kevin Langman <langman@ca.ibm.com>