Added docker file for JDK11 Ubuntu16 x86_64
This docker file can be used to build JDK11 using OpenJ9. It installs
gcc-7.3 and g++7.3 for building JDK11.

Currently, no JDK10 with OpenJ9 releases are available at AdoptOpenJDK.
So, the AdoptOpenJDK API can't be used to download a boot JDK. So, a
nightly JDK10 OpenJ9 build is currently used as the boot JDK. Once JDK10
OpenJ9 releases are available, the following link can be used to
download the JDK10 OpenJ9 boot JDK:
https://api.adoptopenjdk.net/openjdk10-openj9/releases/x64_linux/latest/binary.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>