j.l.Class.getMethods() should select more specific method
As per Java spec:
For methods with same signature (name, parameter types) and return type,
only the most specific method should be selected. Method N is more
specific than M if:
N is declared by a class and M is declared by an interface; or
N and M are both declared by either classes or interfaces and N's
declaring type is the same as or a subtype of M's declaring type.
Select most specific method from a set of methods with same signature
using algorithm above.
Added tests to verify this behaviour.

Signed-off-by: Jason Feng <fengj@ca.ibm.com>