Implement getNewArrayTypeFromClass for JAAS
Given a client primitive array class pointer we need to map it
to the corresponding ordinal for newarray. This is normally done
by via the array of class pointers in J9JavaVM, however for JAAS
the server's J9JavaVM is irrelevant and the mapping has to be
done on the client.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>