Get TR_PersistentJittedBodyInfo from the client
There are two new message types added, one for getting the jitted body
info from a start PC and another for getting it from a resolved method.

Instances where the code grabs the start pc followed immediately by getting
the jitted body info from it have been replaced with a single call to
get the jitted body info from the method, to reduce latency.

Some fields of the jitted body info and method info are ignored and set
to null upon deserialization, including profiling infomation, map tables
and optimization plans. These should never be required on the server,
but in the future we may add support for profiling info as an optimization.

Signed-off-by: Noah Weninger <noah.weninger@ibm.com>