Eliminate remote queries to persistentJittedBodyInfo
Many (if not all) of the queries for jittedBodyInfo are done through the
TR_ResolvedJ9Method interface. The idea implemented in this commit is to
send the required information at the time a TR_ResolvedJ9JAASServerMethod
object is created. The data will be deserialized in heap memory and a pointer
to it will be stored in the newly created TR_ResolvedJ9JAASServerMethod
object.

It should be noted that this bodyInfo describes a body that already exists.
For a method to be compiled the code also creates a bodyInfo using
persistent memory, but this body info is not in play until the compilation
is finished and teh result is sent to the client.
Thus, for a TR_ResolvedJ9Method representing a method undergoing compilation
there could be two bodyInfos at the server
(1) One that describes the previous compilation (uses heap memory)
  We should access this through compilee->getExistingJittedBodyInfo()
(1) Another one that describes current compilation (uses persistent memory)
  We should access this through the recompilation object
  compiler->getRecompilationInfo()->getJittedBodyInfo()

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>