Throw IAE when last char in replacement string is '$' or '\'
The last character of String substitute in j.l.String.replaceAll(regex,
substitute) can't be '$' which is treated as references to captured
subsequences, or '\' which is used to escape literal characters.
'$' requires a group index to follow, and '\' requires a character to be
escaped.
Added tests to verify this behaviour.

Signed-off-by: Jason Feng <fengj@ca.ibm.com>