Generate all DDR pointer types
Whether a generated pointer class is needed is independent of how
many fields or constants it might declare. We could analyze all
the types and generate only those with fields, or that are used
by some other type, but not all references can be expressed in
the superset.dat file. In particular, the format of that file
permits a type to have at most one superclass, and there are C++
classes that employ multiple inheritance (e.g. SH_CompositeCacheImpl
inherits from both SH_CompositeCache and AbstractMemoryPermission).
As a result, a few extra classes will be generated, but they will be
small (they have no fields else they would been generated previously).
On zLinux, a few is about 140.

Minor cleanup:
* remove empty 'omitList' and its uses
* simplify generation of base class name

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>