Move all files from compiler/trj9 to compiler
In addition delete the trj9 directory all together.

*** IMPORTANT ***

If you are using `git log` to look at the history of a particular
file your history will stop at this commit due to movement of all the
files in the compiler directory.

To get the full history use `git log --follow` instead.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>