Fix a debug build failing due to incorrect reference count
In building transaction regions, when we have a monent under
nullchk, we rip out nullchk to new tree and replace it's first
child with newly created PassThrough node. We should increase
reference count of PassThrough node when we set it as first child
of NullChk node. Incorrect reference count was causing asserts to
be thrown.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>