Prefix include directives with "omr/" to explicitly show intention
For an include such as:

```
#include "compiler/z/codegen/TreeEvaluatorTable.hpp"
```

The intention of the include is not very clear. The above include is
supposed to include the OMR version of the tree evaluator table.
However this is not very clear just from the include path since the
same path now exists in OpenJ9. The reason why OMR's version will be
picked is because the /openj9/runtime directory is not included in the
compilers IPATH directives.

Unless someone knows this it would be hard to determine which include
actually gets picked up. As such we modify these confusing include
directives and prefix the omr directory so it is very clear.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>