Added server cache for getRemoteROMString
All of the strings received in TR_ResolvedJ9JAASServerMethod:getRemoteROMString are now cached in a hashtable.
A query to the client is done only if the requested string is not in the cache.

Running `java -XX:jaasClient -Xjit:count=0 -version` with this improvement leads to 85% reduction
in the number of getRemoteROMString queries.

I created a new type of hashtable, the one that takes TR::Region as allocator, instead of TR::PersistentAllocator,
because the hashtable of strings needs to be destroyed at the end of every compilation.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>