Reduce number of resolved methods created in StringPeepholes
Profiles show that creating a TR_ResolvedMethod is quite expensive in terms
of CPU cycles. Many resolved methods are created during StringPeepholes
optimization in findSymRefForValueOf() where the JIT scans the String class
for a specific method identified by name and signature. The exiting code
creates a resolved method for every method in the String class, puts those
resolved methods in a linked list and then starts to do name/signature
comparisons. This is inefficient both in tems of CPU and memory.

This commit changes the algorithm so that the JIT does the name/signature
comparison as it visits each method in the String class. As a result, only
one TR_ResolvedMethod is created and the linked list is eliminated.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>