Don't free remote compile body info
The body info for a remote compile is (currently) in
the data cache and should not be freed via the general
persistent memory free routine. This patch adds a flag
to the body info to indicate that the body is a remote
compile body for this purpose and also for future use.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>