Make java/lang/Class available in AOT
Currently getClassClassPointer() refuses to return the J9Class pointer
for java/lang/Class when compiling AOT (ahead of time) even though Class
is practically guaranteed to be identical at AOT load time.

This commit allows callers to opt in to finding Class, in which case
they must ensure that their use of the J9Class pointer is acceptable for
AOT compilation.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>