Improve entropy of reply key
com.ibm.tools.attach.target.Reply uses a randomly generated key to verify that
the process responding to an attach attempt is the intended target.  The
existing approach used a random value based on time, which makes it
predictable. The key also had only 32 bits of entropy.

Use SecureRandom to generate a longer and more random value when security is
important.

Signed-off-by: Peter Bain <peter_bain@ca.ibm.com>