Update readme for dependent branches
PullRequests support building with dependent changes
in branches now too. This is particularly useful
when you have an OMR change that was merged and
your change is dependent. Because the PR is merged,
you cannot build with the dependent PR because the
merge ref is removed when the PR is merged. The default
branch to build omr is openj9. If you build against
master, you will pickup the merged dependent PR.

Issue #827
Related #1698
[skip ci]

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>