Tenure char[] for interned String objects
Interned String objects are created in tenure space. Currently,
the char[] backing these String objects is left in the nursery.
This means that until the char[] is tenure the Remembered Set has
to contain an entry for each new interned String. To reduce the
overhead on the RS we are going to tenure the char[].

Historically this was happening but the change got lost in
refactoring a few years ago.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>