Fix Escape Analysis's is(NonThis)ArgToCall APIs
EA reserved 8 bits to indicate whether a specific object candidate is
an argument or this argument to a method call at certain depth. 9 bits
are needed when the depths is equal to 8 which is the maximum sniffing
depth. Change the number of bits reserved to 16bits to avoid overflow
when sniffing at maximum depth.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>