Change API used in compilation to keep class alive
Compilation thread needs to keep the object of the class of the
compiling method alive to prevent the class from being unloaded
while compiling. This is achieved by pushing the class object
on to the Java stack before compilation starts and poping
the object after the compilation is finished.
PUSH/POP_OBJECT_IN_SPECIAL_FRAME were used. However, this is
incompatible with the way knowObjectTable works where it pushes
the known objects onto the same Java stack but set the objects
to NULL instead of popping them.

Change PUSH/POP_OBJECT_IN_SPECIAL_FRAME in compilationThread to
j9jni_createLocalRef so that it works correctly with the
knowObjectTable.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>