Fix class redefinition in JITaaS
When jitClassRedefined adds a redefined class to the unload list, incorrect
pointer to the stale class is added.

In J9CompileDispatcher::compile, unloaded classes are processed after
the ROM class cache is accessed, which leads to an invalid class being
accessed when trying to get a newly redefined class. Need to remove
redefined class from the cache first.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>