Make linking similar to linux on osx
Based on suggestions on stackoverflow (see link in code)
I modified the link options to work around a missing reference
to __cxa_pure_virtual.

Running `nm` on libj9shr29.so on mac with this change results in:
```
                 U ___cxa_pure_virtual
```
while it on linux we get:
```
00000000000648f0 t __cxa_pure_virtual
```

This may not work to actually run `java -version` but let's us
find the next set of problems.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>