Improve JProfilingValue Trees
In compilations which use jProfiling, for each value we decide to profile,
we generate a code that does the following.
if ( value == table.keys[hash(value)])
   ++freqs[hash(value)];
else if ( table.otherIndex >= 0 )
   ++freqs[otherIndex];
else
   call _jProfile(32/64)BitValue
In summary we first check if the value exists in hash table and if exists
we increase corresponding frequency counter.
In cases when value does not exist in the hash table,
we need to check if we can acquire lock on table and add new value to it,
other wise we need to increase other frequency.
This commit optimizes above sequence by using ternary select operation to
decide which frequency counter (hash(value) or otherIndex) we need to increment.
With this change, we now only have three blocks per value we profile instead of
six.
Also in case of profiling VFT pointer, we used to check if the object whose J9Class
we are profilig is NULL or not and in case it is NULL, to avoid dereferencing NULL in
profiling code we used to generate a fallBack value block.
There is no need to generate this as we can simply skip the profiling code if object
whose J9Class pointer we are profiling is NULL.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>