z/TPF Specific Build Updates for better integration with z/TPF builds
Today zTPF build does not integrate jilconst and ddrgen
execution, and requires manual intervention to complete the build.
Other platforms are able to run these executables inline with
the rest of the build.  These updates will allow z/TPF to do the
same.

z/TPF is a cross compiled platform, that is to say development and
production environments are very different and what runs on the
development environment will not run on the production environment.
ddrgen and jilconst builds for other platforms assume these systems
to be the same.  As part of this pull request we added the capability
for z/TPF development (openj9 builds) to launch a shell script which
will invoke a REST service to z/TPF which will load the appropriate
programs created in the middle of the build on the development
platform and run them.  Upon completion of the REST service z/TPF will
respond with the generated artifacts and the shell script on
the development system will touch a ".completed" file which will be
used by subsequent openj9 build phases to avoid re-firing the REST
request duplicating the corresponding ddrgen and jilconst
artifacts. The development system continues with its build, as
if the artifacts were generated inline on the same build platform.

Additionally this pull request has miscellaneous updates for both
cleanup and better integration with the z/TPF build environment.  For
example, we updated different path rules and variable labels to match
the rest of our operating system builds.

Signed-off-by: James D Johnston <jjohnst@us.ibm.com>