Correctly identify use of compressed references for CAS write barrier
The `usingCompressedReferences` variable is misleading as it does not
imply compressed references are actually used. What this variable
actually means is that the value we are storing into the object via
compare and swap is a compressed reference.

Recognized CAS sequences are reduced at lower trees phase [1] where the
last two arguments being the expected value and the new value are
lowered into compressed objects since under compressed references
objects on heap are always compressed.

For concurrent scavenge we need a read barrier on the expected value
hence the read barrier under compressed references need not rely on the
use of the `usingCompressedReferences` variable which means something
totally different.

We take this opportunity to first fix the bug and to improve the
meaning of some of the variables in the modified function so it is not
confusing for future readers.

[1] https://github.com/eclipse/openj9/blob/8a36fcd5f23005bf5028d28b60fd61aa325940a1/runtime/compiler/codegen/J9CodeGenerator.cpp#L675-L683

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>