Fail compilation when GC happens under certain conditions
`TR_J9ByteCodeIlGenerator::runFEMacro` for the case of
`TR::java_lang_invoke_ILGenMacros_invokeExactAndFixup` requires
2 remote messages: the first one to dereference a pointer to uintptrj_t,
  which is then used to walk a reference chain on the server,
  and the second one to do things that require VM access on the client.

If GC happens in-between these messages, something
that is accessed with the dereferenced pointer might get garbage
collected, making the second message access invalid data.
The same problem happens in some other cases

Work around this problem by failing a compilation if GC happens
in-between the first and the second messages.

This is done by keeping a global counter of local GC cycles,
and a counter per compilation thread. When the first message
in a sequence is received, per thread counter is updated to the
global value. When the next message is received, we check that
the local counter is still equal to the global counter, ensuring
that no GC occurred.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>