Stop reserving monitors that have been cancelled
Today, a monitor that was reserved and then cancelled may be reserved
again. It is not difficult to accidentally cause an arbitrary number of
cancellations of a single monitor - one need only access a shared
StringBuffer from multiple threads. Due to the high cost of cancellation
(which waits for the thread holding the reservation to yield), lock
reservation can significantly degrade performance when this usage
pattern occurs. Real applications have seen large regressions (17% in
one case) caused by the JIT compiler enabling reservation by default.

This commit implements a scheme to prevent reserving a monitor after it
has been cancelled, as proposed in issue #1124. A monitor whose lock
word is 0 can no longer be reserved. Instead, reservation requires that
the previous value of the lock word contain a particular bit pattern to
signify that the monitor is reservable. The reservable bit pattern has
all bits clear except for the reservation bit, which is set. There was
previously no sensible interpretation for this value in a lock word, so
once a monitor is acquired (either with reservation or without), its
reservable status will not be restored thereafter. Objects subject to
reservation have their lock words initialized to the reservable value in
order to allow them to be reserved once each.

This new scheme is a more conservative default which prevents the
pathological behaviour described above while retaining most of the
benefit of lock reservation for objects that are only ever locked by a
single thread.

In cases where it is known that reserved objects are not shared between
threads, or where it is desirable to allow monitors to be serially
reserved by different threads, users can opt in to the original
behaviour by passing -XlockReservation on the command line.

For the moment, the new scheme is used only on x86. This commit includes
a number of tweaks to x86 code generation in order to accommodate it:

- The reserving entry fast path falls back to lock cmpxchg when the lock
  word is zero, to prevent any attempt to reserve.

- The non-reserving entry fast path accepts reservable locks.

- The "non-reserving" entry fast path now conditionally reserves, if the
  type is known to be one for which reservation is in use.

- The reserving exit fast path falls back to a sequence for unreserved
  locks (cmp ebp, [lock word]; jne; mov [lock word], 0).

- Outlined recursive increment and decrement no longer spuriously reject
  reserved locks.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>